Added Transactional Testing, Workspace Managers and Shared Data between fixtures
Transactional Testing is sharing the same connection between all request. It is up to the tester to begin/rollback transaction in the test.

Workspace Manager is a DAO for workspace, it is made of two layers : managing doctrine and managing ACL. it is available through Symfony Service Container

Testing data made in Fixture can be shared using a reference container, the tester must provide the same container to all the ficture which share data.